refactor: move is_character and string_traits into String.hpp
