+++ include/mimic++/Fwd.hpp
@@ -69,14 +69,6 @@ namespace mimicpp
 	template <typename T>
 	struct is_character;
 
-	/**
-	 * \brief Convenience boolean-constant to the result of ``is_character`` trait.
-	 * \tparam T The type to check.
-	 * \ingroup TYPE_TRAITS_IS_CHARACTER
-	 */
-	template <typename T>
-	inline constexpr bool is_character_v{is_character<T>::value};
-
 	template <typename T>
 	struct string_traits;
